chore: isolate code related to credentials
